+++ index.html
@@ -11,20 +11,24 @@
     <div class="calculator-body">
       <div class="enclosing-calculator">
         <div class="text-box">
+          <!--View History Button-->
           <img
             src="./images/history.png"
             class="imageHW hamburger-menu view-history"
             alt="visit history button"
             tabindex="0"
           />
+          <!--Toggle from day and night mode-->
           <img
             src="./images/brightness-and-contrast.png"
             alt="button for light and dark mode"
             class="imageHW hamburger-menu dark-light-button"
             tabindex="0"
           />
+          <!--Input Text Box-->
           <input type="text" class="calculator-input" placeholder="0" />
         </div>
+        <!--Advanced Operations-->
         <div class="advanced-operations">
           <div class="deg-fe">
             <div class="deg-rad-button" tabindex="0">DEG</div>
@@ -64,6 +68,7 @@
             </div>
           </div>
         </div>
+        <!-- Basic Operations -->
         <div class="calculator-buttons">
           <div
             aria-label="calculator operation buttons"
@@ -71,7 +76,7 @@
             tabindex="0"
             class="2nd-button"
           >
-            <p>2nd</p>
+            <p>2<sup>nd</sup></p>
           </div>
           <div
             aria-label="calculator operation buttons"
